git-svn-id: file:///semplestb/final_repo@1432 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMplestNIOProtocol/src/semplest/services/client/interfaces/SemplestBiddingInterface.java
@@ -8,7 +8,7 @@ import semplest.server.protocol.adengine.KeywordDataObject;
 
 public interface SemplestBiddingInterface extends ServiceInitialize {
 	
-	HashMap<String, Integer> GetMonthlyBudgetPercentPerSE(Integer promotionID, ArrayList<String> searchEngine) throws Exception;
+	HashMap<String, Double> GetMonthlyBudgetPercentPerSE(Integer promotionID, ArrayList<String> searchEngine) throws Exception;
 	HashMap<String, AdEngineInitialData> getInitialValues(Integer promotionID, ArrayList<String> searchEngine) throws Exception;
 	void setBidsInitial(Integer promotionID, String searchEngine) throws Exception;
 	void setBidsUpdate(Integer promotionID, String searchEngine) throws Exception;
